:bug: fix: correct  'res.code( )' to 'res.status( )'

+++ src/services/allocation.service.ts
@@ -1,11 +1,14 @@
 import { FastifyReply, FastifyRequest } from 'fastify'
 import { z } from 'zod'
+import {
+  getAllocationsByIdDetentoModel,
+  getAllocationsByIdModel,
+} from './../models/allocation.model'
 
 import {
   createAllocationsModel,
   deleteAllocationsModel,
   getAllAllocationsModel,
-  getAllocationsByIdModel,
 } from '../models/allocation.model'
 import { getCellByIdwithAlocationsModel } from '../models/cell.model'
 
@@ -24,31 +27,31 @@ export const createAllocationService = async (
     const cela = await getCellByIdwithAlocationsModel(celaId)
 
     if (!cela) {
-      res.code(404).send({ error: 'Cela não encontrada' })
+      res.status(404).send({ error: 'Cela não encontrada' })
       return
     }
 
     // Verificar se a cela já atingiu a capacidade máxima
     if (cela && cela.alocacoes.length >= cela.capacidade) {
       res
-        .code(400)
+        .status(400)
         .send({ error: 'Cela lotada. Não é possível alocar mais detentos.' })
       return
     }
     // Verificar se o detento já está alocado
-    const detentoAlocado = await getAllocationsByIdModel(detentoId)
+    const detentoAlocado = await getAllocationsByIdDetentoModel(detentoId)
 
     if (detentoAlocado) {
-      res.code(400).send({ error: 'Detento já está alocado em uma cela.' })
+      res.status(400).send({ error: 'Detento já está alocado em uma cela.' })
       return
     }
 
     const alocacao = await createAllocationsModel(detentoId, celaId)
 
-    res.code(201).send(alocacao)
+    res.status(201).send(alocacao)
   } catch (error) {
     console.error('Erro ao criar alocação:', error)
-    res.code(500).send({ error: 'Erro interno do servidor' })
+    res.status(500).send({ error: 'Erro interno do servidor' })
     return
   }
 }
@@ -59,10 +62,11 @@ export const getAllAllocationService = async (
 ) => {
   try {
     const alocacoes = await getAllAllocationsModel()
-    res.code(200).send(alocacoes)
+
+    res.status(200).send(alocacoes)
   } catch (error) {
     console.error('Erro ao listar alocações:', error)
-    res.code(500).send({ error: 'Erro interno do servidor' })
+    res.status(500).send({ error: 'Erro interno do servidor' })
     return
   }
 }
@@ -77,11 +81,19 @@ export const deleteAllocationService = async (
   const { id } = validationParams.parse(req.params)
 
   try {
+    const allocationExist = await getAllocationsByIdModel(id)
+
+    if (!allocationExist?.celaId) {
+      res.status(404).send({ error: 'Alocação não encontrada' })
+      return
+    }
+
     await deleteAllocationsModel(id)
-    res.code(200).send({ message: 'Alocação removida com sucesso' })
+
+    res.status(200).send({ message: 'Alocação removida com sucesso' })
   } catch (error) {
     console.error('Erro ao remover alocação:', error)
-    res.code(500).send({ error: 'Erro interno do servidor' })
+    res.status(500).send({ error: 'Erro interno do servidor' })
     return
   }
 }
